gh-620: Explicitly include `InstalledVersions.php` and `installed.php` in PHAR
The autoloadmap includes a mapping `Composer\InstalledVersions` but as
of explicitly including `src` this file is no longer included.

This commit explicitly adds both `vendor/composer/InstalledVersions.php`
and `vendor/composer/installed.php` to ensure that the previous behavior
is preserved and WARNINGS are not issued when other packages use
`InstalledVersions::getVersion(// ...)`.